feat(sse): change to redux-toolkit actions
