+++ src/modules/oh-indicators-list/store/index.ts
@@ -1,6 +1,6 @@
 import reducer from './slice';
 
-export { selectItems, selectErrorMessage } from './selectors';
-export { getAllItems } from './actions';
+export * as selectors from './selectors';
+export * as actions from './actions';
 export { indicatorsSlice } from './slice';
 export { reducer as default };
