+++ src/modules/oh-indicators-list/store/middlewares.ts
@@ -5,16 +5,6 @@ import {
 } from '../network-layer/sseClient';
 import { connectionClosed, connectionOpened, receiveMessage } from './actions';
 
-// todo: move to a separate place
-const parseStreamingResponse = (
-  incomingMessage: MessageEvent,
-): { name: string; value: any } => {
-  const data = JSON.parse(incomingMessage.data);
-  const [, , name] = data.topic.split('/');
-  const { value } = JSON.parse(data.payload);
-  return { name, value };
-};
-
 // todo: types
 export const sseMiddleware = (
   client: IEventSourceInitializer & IEventSourceFinalizer,
@@ -27,9 +17,7 @@ export const sseMiddleware = (
         onMessageHandler: (event) => {
           const data = JSON.parse(event.data);
           if (data.type === 'ItemStateChangedEvent') {
-            return store.dispatch(
-              receiveMessage(parseStreamingResponse(event)),
-            );
+            return store.dispatch(receiveMessage!(event));
           }
         },
         onOpenHandler: () => store.dispatch(connectionOpened()),
